FASTQ Input instead of SRA Accessions

+++ scripts/cleave.sh
@@ -1,7 +1,18 @@
 #!/usr/bin/env bash
 set -euo pipefail
-PROGRAM=$(basename $0)
+FULL_PROGRAM=$0
+PROGRAM=$(basename $FULL_PROGRAM)
+args="$FULL_PROGRAM $*"
 
+# 0 - table function
+function table() {
+	if column -L <(echo) &>/dev/null; then
+		cat | column -s $'\t' -t -L
+	else
+		cat | column -s $'\t' -t
+		echo
+	fi
+}
 # 1 - get_help function
 function get_help() {
 	{
@@ -15,6 +26,7 @@ function get_help() {
 		\t  - cleaved.mature.len.faa\n \
 		\t  - CLEAVE.DONE or CLEAVE.FAIL\n \
 		\t  - CLEAVE_LEN.DONE or CLEAVE_LEN.FAIL\n \
+		\t  - CLEAVE_LEN_NR.DONE or CLEAVE_LEN_NR.FAIL\n \
 		\n \
 		\tEXIT CODES:\n \
 		\t-----------\n \
@@ -23,27 +35,30 @@ function get_help() {
 		\t  - 2: SignalP not found\n \
 		\t  - 3: cleavage failed\n \
 		\t  - 4: length filtering failed\n \
+		\t  - 5: redundancy removal failed\n \
 		\n \
 		\tFor more information on ProP: https://services.healthtech.dtu.dk/service.php?ProP-1.0\n \
-		" | column -s $'\t' -t -L
+		" | table
 
 		echo "USAGE(S):"
 		echo -e "\
-		\t$PROGRAM [OPTIONS] -o <output directory> <input FASTA file>\n \
-		" | column -s $'\t' -t -L
+		\t$PROGRAM [-a <address>] [-c] [-d] [-h] [-s <0 to 1>] -o <output directory> <input FASTA file>\n \
+		" | table
 
 		echo "OPTION(S):"
 		echo -e "\
-		\t-a <address>\temail alert\n \
+		\t-a <address>\temail address for alerts\n \
 		\t-c\tallow consecutive (i.e. adjacent) segments to be recombined\n \
+		\t-d\tdebug mode (skips running ProP)\n \
 		\t-h\tshow this help menu\n \
 		\t-o <directory>\toutput directory\t(required)\n \
-		" | column -s $'\t' -t -L
+		\t-s <0 to 1>\tredundancy removal cut-off\t(default = 1.0)\n \
+		" | table
 
 		echo "EXAMPLE(S):"
 		echo -e "\
-		\t$PROGRAM -o /path/to/cleavage /path/to/homology/jackhmmer.nr.faa\n \
-		" | column -s $'\t' -t -L
+		\t$PROGRAM -a user@example.com -c -o /path/to/cleavage/outdir /path/to/homology/jackhmmer.nr.faa\n \
+		" | table
 	} 1>&2
 	exit 1
 
@@ -51,13 +66,25 @@ function get_help() {
 	#		\t-s <0 to 1>\tCD-HIT global sequence similarity cut-off\t(default = 0.90)\n \
 }
 
+# 1.5 - print_line function
+function print_line() {
+	if command -v tput &>/dev/null; then
+		end=$(tput cols)
+	else
+		end=50
+	fi
+	{
+		printf '%.0s=' $(seq 1 $end)
+		echo
+	} 1>&2
+}
 # 2 - print_error function
 function print_error() {
 	{
+		echo -e "CALL: $args (wd: $(pwd))\n"
 		message="$1"
 		echo "ERROR: $message"
-		printf '%.0s=' $(seq 1 $(tput cols))
-		echo
+		print_line
 		get_help
 	} 1>&2
 }
@@ -72,8 +99,10 @@ email=false
 # similarity=0.90
 consecutive=false
 outdir=""
+similarity=1.0
+debug=false
 # 4 - read options
-while getopts :a:cho: opt; do
+while getopts :a:cho:s:d opt; do
 	case $opt in
 	a)
 		address="$OPTARG"
@@ -82,11 +111,14 @@ while getopts :a:cho: opt; do
 	c)
 		consecutive=true
 		;;
+	d)
+		debug=true
+		;;
 	h) get_help ;;
 	o)
 		outdir="$(realpath $OPTARG)"
 		;;
-		#		s) similarity="$OPTARG";;
+	s) similarity="$OPTARG" ;;
 	\?)
 		print_error "Invalid option: -$OPTARG" 1>&2
 		;;
@@ -109,7 +141,11 @@ fi
 if [[ ! -f $(realpath $1) ]]; then
 	print_error "Input file $(realpath $1) does not exist."
 elif [[ ! -s $(realpath $1) ]]; then
-	print_error "input file $(realpath $1) is empty."
+	print_error "Input file $(realpath $1) is empty."
+fi
+
+if (($(echo "$similarity <= 0" | bc -l) || $(echo "$similarity > 1" | bc -l))); then
+	print_error "Invalid argument for -r <0 to 1>: $similarity"
 fi
 
 # 7 - remove existing status files
@@ -117,13 +153,18 @@ rm -f $outdir/CLEAVE.DONE
 rm -f $outdir/CLEAVE.FAIL
 rm -f $outdir/CLEAVE_LEN.DONE
 rm -f $outdir/CLEAVE_LEN.FAIL
+rm -f $outdir/CLEAVE_LEN_NR.DONE
+rm -f $outdir/CLEAVE_LEN_NR.FAIL
 
 # 8 - print env details
-echo "HOSTNAME: $(hostname)" 1>&2
-echo -e "START: $(date)\n" 1>&2
-# start_sec=$(date '+%s')
+{
+	echo "HOSTNAME: $(hostname)"
+	echo -e "START: $(date)\n"
 
-echo -e "PATH=$PATH\n" 1>&2
+	echo -e "PATH=$PATH\n"
+
+	echo -e "CALL: $args (wd: $(pwd))\n"
+} 1>&2
 
 if ! command -v mail &>/dev/null; then
 	email=false
@@ -133,97 +174,111 @@ fi
 infile=$(realpath $1)
 tempfile=$outdir/prop.out
 
+if [[ ! -v WORKDIR ]]; then
+	workdir=$(dirname $outdir)
+else
+	workdir=$(realpath $WORKDIR)
+fi
+
+if [[ ! -v SPECIES ]]; then
+	# get species from workdir
+	species=$(echo "$workdir" | awk -F "/" '{print $(NF-1)}')
+else
+	species=$SPECIES
+fi
+
 echo "PROGRAM: $(command -v $RUN_PROP)" 1>&2
 echo -e "VERSION: 1.0c\n" 1>&2
-propdir=$(dirname $RUN_PROP)
+# propdir=$(dirname $RUN_PROP)
 
-if [[ -f "$propdir/CONFIG.DONE" ]]; then
-	config=false
-	echo -e "ProP and SignalP have been pre-configured. Skipping configuration. If this is not the case, please delete the $propdir/CONFIG.DONE file to trigger a reconfiguration of ProP and SignalP.\n" 1>&2
-else
-	config=true
-fi
+# check if this needs to be checked
+# if [[ -f "$propdir/CONFIG.DONE" ]]; then
+# 	config=false
+# 	echo -e "ProP and SignalP have been pre-configured. Skipping configuration. If this is not the case, please delete the $propdir/CONFIG.DONE file to trigger a reconfiguration of ProP and SignalP.\n" 1>&2
+# else
+# 	config=true
+# fi
 
-if [[ "$config" == true ]]; then
-	permissions=$(ls -ld $propdir/tmp | awk '{print $1}')
-	owner=$(ls -ld $propdir/tmp | awk '{print $3}')
-	if [[ "$permissions" != "drwxrw[sx]rwt" && "$owner" == "$(whoami)" ]]; then
-		chmod 1777 $propdir/tmp
-	fi
-fi
+# if [[ "$config" == true ]]; then
+# 	permissions=$(ls -ld $propdir/tmp | awk '{print $1}')
+# 	owner=$(ls -ld $propdir/tmp | awk '{print $3}')
+# 	if [[ "$permissions" != "drwxrw[sx]rwt" && "$owner" == "$(whoami)" ]]; then
+# 		chmod 1777 $propdir/tmp
+# 	fi
+# fi
 # echo -e "VERSION: $(echo $RUN_PROP | grep -o prop\-[0-9]\.[0-9]. | cut -f2 -d-)\n" 1>&2
+# this should have been moved to config-signalp.sh
+# if command -v $RUN_SIGNALP &>/dev/null; then
+# 	signalp=true
+# 	signalp_opt="-s"
+# 	echo -e "SignalP program detected. Proceeding with SignalP.\n" 1>&2
+# 	if [[ "$config" = true ]]; then
+# 		signal_dir=$(dirname $RUN_SIGNALP)
+# 		if [[ ! -f "$signal_dir/CONFIG.DONE" ]]; then
+# 			sed -i "s|^SIGNALP=.*|SIGNALP=$signal_dir|" $RUN_SIGNALP
+# 			sed -i "s|^SH=.*|SH=$SHELL|" $RUN_SIGNALP
+# 			permissions=$(ls -ld $signal_dir/tmp | awk '{print $1}')
+# 			owner=$(ls -ld $signal_dir/tmp | awk '{print $3}')
+# 			if [[ "$permissions" != "drwxrw[sx]rwt" && "$owner" == "$(whoami)" ]]; then
+# 				chmod 1777 $signal_dir/tmp
+# 			fi
+# 			touch $signal_dir/CONFIG.DONE
+# 		else
+# 			echo -e "SignalP has been previously configured.\n" 1>&2
+# 		fi
+#
+# 	fi
+# else
+# 	signalp=false
+# 	signalp_opt=""
+# 	echo "ERROR: SignalP program not found. Please download SignalP into $ROOT_DIR/src, and source $ROOT_DIR/scripts/config.sh from the $ROOT_DIR, so that the RUN_SIGNALP environment variable is re-exported." 1>&2
+# 	if [[ "$email" = true ]]; then
+# 		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
+# 		# echo "$outdir" | mail -s "Failed cleaving peptides for $org" $address
+# 		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
+# 		echo "$outdir" | mail -s "${species^}: STAGE 09: CLEAVAGE: FAILED" $address
+# 		echo "Email alert sent to $address." 1>&2
+# 	fi
+# 	exit 2
+# #	echo -e "SignalP program not found. Proceeding without SignalP.\n" 1>&2
+# fi
+# if [[ "$signalp" = true ]]; then
+echo "PROGRAM: $(command -v $RUN_SIGNALP)" 1>&2
+echo -e "VERSION: $($RUN_SIGNALP -v)\n" 1>&2
+# fi
 
-if command -v $RUN_SIGNALP &>/dev/null; then
-	signalp=true
-	signalp_opt="-s"
-	echo -e "SignalP program detected. Proceeding with SignalP.\n" 1>&2
-	if [[ "$config" = true ]]; then
-		signal_dir=$(dirname $RUN_SIGNALP)
-		if [[ ! -f "$signal_dir/CONFIG.DONE" ]]; then
-			sed -i "s|^SIGNALP=.*|SIGNALP=$signal_dir|" $RUN_SIGNALP
-			sed -i "s|^SH=.*|SH=$SHELL|" $RUN_SIGNALP
-			permissions=$(ls -ld $signal_dir/tmp | awk '{print $1}')
-			owner=$(ls -ld $signal_dir/tmp | awk '{print $3}')
-			if [[ "$permissions" != "drwxrw[sx]rwt" && "$owner" == "$(whoami)" ]]; then
-				chmod 1777 $signal_dir/tmp
-			fi
-			touch $signal_dir/CONFIG.DONE
-		else
-			echo -e "SignalP has been previously configured.\n" 1>&2
-		fi
-
-	fi
-else
-	signalp=false
-	signalp_opt=""
-	echo "ERROR: SignalP program not found. Please download SignalP into $ROOT_DIR/src, and source $ROOT_DIR/scripts/config.sh from the $ROOT_DIR, so that the RUN_SIGNALP environment variable is re-exported." 1>&2
-	if [[ "$email" = true ]]; then
-		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
-		# echo "$outdir" | mail -s "Failed cleaving peptides for $org" $address
-		org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
-		echo "$outdir" | mail -s "${org^}: STAGE 09: CLEAVAGE: FAILED" $address
-		echo "Email alert sent to $address." 1>&2
-	fi
-	exit 2
-#	echo -e "SignalP program not found. Proceeding without SignalP.\n" 1>&2
-fi
-if [[ "$signalp" = true ]]; then
-	echo "PROGRAM: $(command -v $RUN_SIGNALP)" 1>&2
-	echo -e "VERSION: $($RUN_SIGNALP -v)\n" 1>&2
-fi
-
-if [[ "$config" = true ]]; then
-	if [[ ! -f $propdir/CONFIG.DONE ]]; then
-		echo -e "Configuring ProP...\n" 1>&2
-		sed -i "s|setenv\tPROPHOME.*|setenv\tPROPHOME\t$propdir|" $RUN_PROP
-
-		awkbin=$(command -v awk)
-		sed -i "s|setenv AWK.*|setenv AWK $awkbin|" $RUN_PROP
-		sed -i 's/^AWK=.*/AWK=awk/' $RUN_SIGNALP
-		sed -i "s|AWK=/.*|AWK=$awkbin|" $RUN_SIGNALP
-
-		echobin=$(which echo)
-		sed -i "s|setenv ECHO.*|setenv ECHO \"$echobin -e\"|" $RUN_PROP
-
-		gnuplot=$(command -v gnuplot 2>/dev/null || true)
-		if [[ ! -z $gnuplot ]]; then
-			sed -i "s|setenv GNUPLOT.*|setenv GNUPLOT $gnuplot|" $RUN_PROP
-			sed -i "s|PLOTTER=/.*|PLOTTER=$gnuplot|" $RUN_SIGNALP
-		fi
-
-		ppmtogifbin=$(command -v ppmtogif 2>/dev/null || true)
-		if [[ ! -z $ppmtogifbin ]]; then
-			sed -i "s|setenv PPM2GIF.*|setenv PPM2GIF $ppmtogifbin|" $RUN_PROP
-			sed -i "s|PPMTOGIF=/.*|PPMTOGIF=$ppmtogifbin|" $RUN_SIGNALP
-		fi
-
-		if [[ "$signalp" = true ]]; then
-			sed -i "s|setenv SIGNALP.*|setenv SIGNALP $RUN_SIGNALP|" $RUN_PROP
-		fi
-	else
-		echo -e "ProP has been previously configured.\n" 1>&2
-	fi
-fi
+# if [[ "$config" = true ]]; then
+# 	if [[ ! -f $propdir/CONFIG.DONE ]]; then
+# 		echo -e "Configuring ProP...\n" 1>&2
+# 		sed -i "s|setenv\tPROPHOME.*|setenv\tPROPHOME\t$propdir|" $RUN_PROP
+#
+# 		awkbin=$(command -v awk)
+# 		sed -i "s|setenv AWK.*|setenv AWK $awkbin|" $RUN_PROP
+# 		sed -i 's/^AWK=.*/AWK=awk/' $RUN_SIGNALP
+# 		sed -i "s|AWK=/.*|AWK=$awkbin|" $RUN_SIGNALP
+#
+# 		echobin=$(which echo)
+# 		sed -i "s|setenv ECHO.*|setenv ECHO \"$echobin -e\"|" $RUN_PROP
+#
+# 		gnuplot=$(command -v gnuplot 2>/dev/null || true)
+# 		if [[ ! -z $gnuplot ]]; then
+# 			sed -i "s|setenv GNUPLOT.*|setenv GNUPLOT $gnuplot|" $RUN_PROP
+# 			sed -i "s|PLOTTER=/.*|PLOTTER=$gnuplot|" $RUN_SIGNALP
+# 		fi
+#
+# 		ppmtogifbin=$(command -v ppmtogif 2>/dev/null || true)
+# 		if [[ ! -z $ppmtogifbin ]]; then
+# 			sed -i "s|setenv PPM2GIF.*|setenv PPM2GIF $ppmtogifbin|" $RUN_PROP
+# 			sed -i "s|PPMTOGIF=/.*|PPMTOGIF=$ppmtogifbin|" $RUN_SIGNALP
+# 		fi
+#
+# 		if [[ "$signalp" = true ]]; then
+# 			sed -i "s|setenv SIGNALP.*|setenv SIGNALP $RUN_SIGNALP|" $RUN_PROP
+# 		fi
+# 	else
+# 		echo -e "ProP has been previously configured.\n" 1>&2
+# 	fi
+# fi
 
 if [[ "$(grep -c "|" $infile)" -gt 0 ]]; then
 	echo -e "NOTE: Pipes detected in sequence headers will be converted to underscores for ProP.\n" 1>&2
@@ -232,17 +287,14 @@ fi
 
 # RUN PROP and get output
 echo "Predicting cleavage sites..." 1>&2
-echo "COMMAND: $RUN_PROP -p $signalp_opt $infile > $tempfile" 1>&2
-# start_sec_temp=$(date '+%s')
+echo "COMMAND: $RUN_PROP -p -s $infile > $tempfile" 1>&2
 
-$RUN_PROP -p $signalp_opt $infile >$tempfile
-# end_sec_temp=$(date '+%s')
-
-cp $tempfile $outdir/prop.raw.out
-sed -i 's/ \+$//' $tempfile
-sed -i 's/^[[:space:]]*[0-9]\+[[:space:]]*/Sequence: /' $tempfile
-
-# $ROOT_DIR/scripts/get-runtime.sh $start_sec_temp $end_sec_temp 1>&2
+if [[ "$debug" = false ]]; then
+	$RUN_PROP -p -s $infile >$tempfile
+	cp $tempfile $outdir/prop.raw.out
+	sed -i 's/ \+$//' $tempfile
+	sed -i 's/^[[:space:]]*[0-9]\+[[:space:]]*/Sequence: /' $tempfile
+fi
 # echo 1>&2
 echo -e "Output: $tempfile\n" 1>&2
 
@@ -252,37 +304,42 @@ echo "Writing ProP results into a separate file for each sequence..." 1>&2
 echo -e "COMAMND: awk -v var=\"$outdir\" 'BEGIN{x=\"/dev/null\"}/^Sequence:/{x=var\"/F\"++i\".txt\";}{print > x;}' $tempfile\n" 1>&2
 # echo -e "COMAMND: awk -v var=\"$outdir\" 'BEGIN{x=\"/dev/null\"}/^\\t[0-9]+/{x=var\"/F\"++i\".txt\";}{print > x;}' $tempfile\n" 1>&2
 # awk -v var="$outdir" 'BEGIN{x="/dev/null"}/^\t[0-9]+/{x=var"/F"++i".txt";}{print > x;}' $tempfile
-awk -v var="$outdir" 'BEGIN{x="/dev/null"}/^Sequence:/{x=var"/F"++i".txt";}{print > x;}' $tempfile
+if [[ "$debug" = false ]]; then
+	awk -v var="$outdir" 'BEGIN{x="/dev/null"}/^Sequence:/{x=var"/F"++i".txt";}{print > x;}' $tempfile
+fi
 # exit 0
 echo "Converting ProP output to a TSV file..." 1>&2
 tsv=$outdir/prop.tsv
 
-echo -e "Sequence\tSignal Peptide\tPropeptide Cleavage" >$tsv
-for i in $outdir/F*.txt; do
-	seqname=$(head -n1 $i | awk '{print $NF}')
-	signal_site=$(awk '/Signal peptide cleavage site predicted/ {print $NF}' $i)
+if [[ "$debug" = false ]]; then
+	echo -e "Sequence\tSignal Peptide\tPropeptide Cleavage" >$tsv
+	for i in $outdir/F*.txt; do
+		seqname=$(head -n1 $i | awk '{print $NF}')
+		signal_site=$(awk '/Signal peptide cleavage site predicted/ {print $NF}' $i)
 
-	if [[ "$signal_site" == "none" ]]; then # || -z "$signal_site" ]] for if -s isn't used for prop
-		signal_site=0
-	else
-		signal_site=$(awk '/Signal peptide cleavage site predicted/ {print $(NF-3)}' $i)
-	fi
+		if [[ "$signal_site" == "none" ]]; then # || -z "$signal_site" ]] for if -s isn't used for prop
+			signal_site=0
+		else
+			signal_site=$(awk '/Signal peptide cleavage site predicted/ {print $(NF-3)}' $i)
+		fi
 
-	prop_sites=$(awk '/\*ProP\*/ {print $2}' $i | tr '\n' ',' | sed 's/,$//')
+		prop_sites=$(awk '/\*ProP\*/ {print $2}' $i | tr '\n' ',' | sed 's/,$//')
 
-	if [[ -z "$prop_sites" ]]; then
-		prop_sites=0
-	fi
+		if [[ -z "$prop_sites" ]]; then
+			prop_sites=0
+		fi
 
-	echo -e "$seqname\t$signal_site\t$prop_sites" >>$tsv
-	rm $i
-done
-
+		echo -e "$seqname\t$signal_site\t$prop_sites" >>$tsv
+		rm $i
+	done
+fi
 # Sambina's cleaving script
 echo "Cleaving peptides..." 1>&2
 # start_cleave=$(date '+%s')
 echo -e "COMMAND: $ROOT_DIR/scripts/cleave-seq.py $infile $tsv $outdir\n" 1>&2
-$ROOT_DIR/scripts/cleave-seq.py $infile $tsv $outdir
+if [[ "$debug" = false ]]; then
+	$ROOT_DIR/scripts/cleave-seq.py $infile $tsv $outdir
+fi
 # end_cleave=$(date '+%s')
 # $ROOT_DIR/scripts/get-runtime.sh $start_cleave $end_cleave
 # echo 1>&2
@@ -292,14 +349,13 @@ echo "Output Files:" 1>&2
 echo -e "\
 - signal_seq.faa: contains all the signal sequences\n\
 - adjacent_seq.faa: contains all the recombined peptide sequences that have adjacent cleaved sequences\n\
-- mature_cleaved_seq.FASTA: contains all the cleaved sequences\n\
+- mature_cleaved_seq.faa: contains all the cleaved sequences\n\
 \t- includes all mature, prop and prepro sequences\n\
-- recombined_seq.FASTA: Contains all the non-adjacent recombined sequences, both two and three cleaved sequences stitched together\n\
+- recombined_seq.faa: Contains all the non-adjacent recombined sequences, both two and three cleaved sequences stitched together\n\
 \t- includes all mature, pro, and prepro sequences)\n\
 " 1>&2
 
 echo "Combining mature_cleaved_seq.faa and recombined_seq.faa into cleaved.mature.faa..." 1>&2
-
 if [[ "$consecutive" = true ]]; then
 	cat $outdir/mature_cleaved_seq.faa $outdir/recombined_seq.faa $outdir/adjacent_seq.faa >$outdir/cleaved.mature.faa
 else
@@ -314,8 +370,8 @@ if [[ ! -s "$outfile" ]]; then
 	if [[ "$email" = true ]]; then
 		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
 		# echo "$outdir" | mail -s "Failed cleaving peptides for $org" $address
-		org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
-		echo "$outdir" | mail -s "${org^}: STAGE 09: CLEAVAGE: FAILED" $address
+		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
+		echo "$outdir" | mail -s "${species^}: STAGE 09: CLEAVAGE: FAILED" $address
 		echo "Email alert sent to $address." 1>&2
 	fi
 	exit 3
@@ -330,24 +386,50 @@ echo -e "VERSION: $(echo "$seqtk_version" | awk '/Version:/ {print $NF}')\n" 1>&
 echo "Removing sequences with length < 2 or > 200 amino acids..." 1>&2
 echo -e "COMMAND: $RUN_SEQTK subseq $outfile <($RUN_SEQTK comp $outfile | awk '{if(\$2>=2 && \$2<=200) print \$1}') > $outfile_len\n" 1>&2
 $RUN_SEQTK subseq $outfile <($RUN_SEQTK comp $outfile | awk '{if($2>=2 && $2<=200) print $1}') >$outfile_len
-echo -e "Removed $($RUN_SEQTK comp $outfile | awk '{if($2<2 || $2>200) print $1}' | wc -l) sequences.\n" 1>&2
-
-echo -e "Number of sequences remaining: $(grep -c '^>' $outfile_len || true)\n" 1>&2
-
-echo -e "Output: $outfile_len\n" 1>&2
+remove_seqs=$($RUN_SEQTK comp $outfile | awk '{if($2<2 || $2>200) print $1}' | wc -l)
+echo -e "Removed $(printf "%'d" $remove_seqs) sequences due to length.\n" 1>&2
 
 if [[ ! -s $outfile_len ]]; then
 	touch $outdir/CLEAVE_LEN.FAIL
 	echo "ERROR: Length filtering output file $outfile_len does not exist or is empty." 1>&2
 	if [[ "$email" = true ]]; then
 		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
-		org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
-		echo "$outdir" | mail -s "${org^}: STAGE 09: CLEAVAGE: SUCCESS" $address
+		# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
+		echo "$outdir" | mail -s "${species^}: STAGE 09: CLEAVAGE: SUCCESS" $address
 		# echo "$outdir" | mail -s "Failed filtering out long sequences for $org" $address
 		echo "Email alert sent to $address." 1>&2
 	fi
 	exit 4
+else
+	touch $outdir/CLEAVE_LEN.DONE
 fi
+
+echo "Removing duplicate sequences..." 1>&2
+outfile_len_nr=$outdir/cleaved.mature.len.nr.faa
+$ROOT_DIR/scripts/run-cdhit.sh -s $similarity -o ${outfile_len_nr} ${outfile_len}
+
+if [[ ! -s ${outfile_len_nr} ]]; then
+	touch $outdir/CLEAVE_LEN_NR.FAIL
+	echo "ERROR: Redundancy removal output file $outfile_len_nr does not exist or is empty." 1>&2
+	if [[ "$email" = true ]]; then
+		species=$(echo "$species" | sed 's/^./\u&. /')
+		echo "$outdir" | mail -s "${species}: STAGE 09: CLEAVAGE: FAILED" $address
+		echo -e "\nEmail alert sent to $address." 1>&2
+	fi
+	exit 5
+else
+	touch $outdir/CLEAVE_LEN_NR.DONE
+fi
+
+num_len=$(grep -c '^>' $outfile_len || true)
+num_nr=$(grep -c '^>' $outfile_len_nr || true)
+
+remove_seqs=$(echo "$num_len - $num_nr" | bc)
+echo -e "\nRemoved $(printf "%'d" $remove_seqs) duplicate sequences.\n" 1>&2
+
+echo -e "Number of sequences remaining: $(printf "%'d" $num_nr)\n" 1>&2
+
+# echo -e "Output: $outfile_len_nr\n" 1>&2
 
 default_name="$(realpath -s $(dirname $outdir)/cleavage)"
 if [[ "$default_name" != "$outdir" ]]; then
@@ -374,18 +456,18 @@ if [[ "$default_name" != "$outdir" ]]; then
 fi
 
 echo -e "END: $(date)\n" 1>&2
-# end_sec=$(date '+%s')
 
-# $ROOT_DIR/scripts/get-runtime.sh -T $start_sec $end_sec 1>&2
 # echo 1>&2
 
-touch $outdir/CLEAVE_LEN.DONE
-echo "STATUS: DONE." 1>&2
+echo -e "STATUS: DONE.\n" 1>&2
 
+echo "Output: $outfile_len_nr" 1>&2
 if [[ "$email" = true ]]; then
 	# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2), $(NF-1)}')
 	# echo "$outdir" | mail -s "Finished cleaving peptides for $org" $address
-	org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
-	echo "$outdir" | mail -s "${org^}: STAGE 09: CLEAVAGE: SUCCESS" $address
+	# org=$(echo "$outdir" | awk -F "/" '{print $(NF-2)}' | sed 's/^./&. /')
+	species=$(echo "$species" | sed 's/^./\u&. /')
+	# echo "$outdir" | mail -s "${species^}: STAGE 09: CLEAVAGE: SUCCESS" $address
+	echo "$outdir" | mail -s "${species}: STAGE 09: CLEAVAGE: SUCCESS" $address
 	echo -e "\nEmail alert sent to $address." 1>&2
 fi
